+++ projects/angular2-smart-table/src/lib/components/thead/thead.module.ts
@@ -9,7 +9,6 @@ import {NgxSmartTableTheadComponent} from './thead.component';
 import {TheadCreateCancelComponent} from './cells/create-cancel.component';
 import {ActionsTitleComponent} from './cells/actions-title.component';
 import {AddButtonComponent} from './cells/add-button.component';
-import {CheckboxSelectAllComponent} from './cells/checkbox-select-all.component';
 import {ColumnTitleComponent} from './cells/column-title.component';
 import {TitleComponent} from './cells/title/title.component';
 import {TheadFitlersRowComponent} from './rows/thead-filters-row.component';
@@ -22,7 +21,6 @@ const THEAD_COMPONENTS = [
   TheadCreateCancelComponent,
   ActionsTitleComponent,
   AddButtonComponent,
-  CheckboxSelectAllComponent,
   ColumnTitleComponent,
   TitleComponent,
   TheadFitlersRowComponent,