+++ projects/angular2-smart-table/src/public-api.ts
@@ -11,6 +11,7 @@ export { Row } from './lib/lib/data-set/row';
 export { DataSet } from './lib/lib/data-set/data-set';
 export { LocalDataSource } from './lib/lib/data-source/local/local.data-source';
 export { ServerDataSource } from './lib/lib/data-source/server/server.data-source';
+export * from './lib/lib/data-source/data-source'
 export * from './lib/angular2-smart-table.component';
 export * from './lib/lib/settings';
 export * from './lib/lib/events';