+++ CHANGELOG.md
@@ -9,6 +9,12 @@ This document lists the changes introduced by this fork.
   When this is set to `true`, the dropdown will really work like a selection
   instead of a set of predefined search queries. For backwards compatibility,
   the default of this value is `false`.
+* Add proper typing for DataSourceChangeEvent
+* Add new selection mode `multi_filtered` where the "select all" checkbox selects only
+  the elements that match the current filter instead of all elements in the data source
+* Add some DataSource methods regarding "select all" to the documentation
+* Remove undocumented 'all' option for `perPage` because that is incompatible with `number`
+* Fixes that types in data-source.ts were not part of public API
 * Fixes several incorrect links to the sources of the demo examples
 * Fixes wrong documentation about the available filter types